+++ backend/libraries/types/src/canister_wasm.rs
@@ -1,4 +1,4 @@
-use crate::Version;
+use crate::{CanisterId, Version};
 use candid::CandidType;
 use human_readable::ToHumanReadable;
 use serde::{Deserialize, Serialize};
@@ -8,6 +8,8 @@ use std::fmt::{Debug, Formatter};
 #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct UpgradeCanisterWasmArgs {
     pub wasm: CanisterWasm,
+    pub filter: Option<UpgradesFilter>,
+    pub use_for_new_canisters: Option<bool>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Clone)]
@@ -33,6 +35,12 @@ impl Debug for CanisterWasm {
             .field("byte_length", &self.module.len())
             .finish()
     }
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug, Default)]
+pub struct UpgradesFilter {
+    pub include: Vec<CanisterId>,
+    pub exclude: Vec<CanisterId>,
 }
 
 #[derive(Serialize)]