+++ sns/scripts/proposals/build_upgrade_canister_proposal/src/lib.rs
@@ -48,6 +48,8 @@ fn create_proposal(config: Config) -> Result<Proposal, Box<dyn Error>> {
             version: config.version,
             module: wasm_module,
         },
+        filter: None,
+        use_for_new_canisters: None,
     };
 
     let payload = Encode!(&args)?;